restructuring to separate field and particle solver. unfinished
