+++ include/Collective.h
@@ -136,7 +136,11 @@ class Collective
     string getSimName()const{ return (SimName); }
     string getWriteMethod()const{ return (wmethod); }
     string getPoissonCorrection()const{ return (PoissonCorrection); }
+    // the number of the first cycle that is executed
+    int get_first_cycle()const{ return (last_cycle+1); }
+    // the number before the first cycle executed
     int getLast_cycle()const{ return (last_cycle); }
+    int get_final_cycle()const{ return ncycles + get_last_cycle();}
     double getVinj()const{ return (Vinj); }
     double getCGtol()const{ return (CGtol); }
     double getGMREStol()const{ return (GMREStol); }
