+++ include/OutputWrapperFPP.h
@@ -3,8 +3,8 @@
 // ===
 // OutputWrapperFPP: output wrapper for file-per-process output
 //
-//   This class should provide a mechanism to avoid having
-//   repeatedly opening and closing the same file.
+//   This class should provide a mechanism to avoid 
+//   repeated opening and closing of the same file.
 // ===
 #include "ipicfwd.h"
 #include "PSKOutput.h"
@@ -15,6 +15,10 @@ using namespace PSK;
 class OutputWrapperFPP
 {
  private:
+  Collective    *col,
+  VCtopology3D  *vct,
+  Grid3DCU      *grid,
+ private:
   #ifndef NO_HDF5
   PSK::OutputManager < PSK::OutputAdaptor > output_mgr; // Create an Output Manager
   myOutputAgent < PSK::HDF5OutputAdaptor > hdf5_agent;  // Create an Output Agent for HDF5 output
@@ -24,13 +28,17 @@ class OutputWrapperFPP
   string RestartDirName;
   string output_file;
  public:
+  OutputWrapperFPP(
+    Collective    *col_,
+    VCtopology3D  *vct_,
+    Grid3DCU      *grid_)
+  : col(col_),
+    vct(vct_),
+    grid(grid_)
+  {}
   void init_output_files(
-    Collective    *col,
-    VCtopology3D  *vct,
-    Grid3DCU      *grid,
     EMfields3D    *EMf,
-    Particles3D   *part,
-    int ns);
+    Particles3D   *part);
   void append_output(const char* tag, int cycle);
   void append_output(const char* tag, int cycle, int sample);
 };
