+++ main/Kinetics.cpp
@@ -1,0 +1,144 @@
+
+// === Section: Kinetics_routines ===
+
+Kinetics::~Kinetics()
+{
+  // delete particles
+  //
+  if(part)
+  {
+    for (int i = 0; i < ns; i++)
+    {
+      // placement delete
+      part[i].~Particles3D();
+    }
+    free(part);
+  }
+}
+
+Kinetics::Kinetics(const Setting& setting_, Imoments* iMoments_)
+: setting(setting_),
+  iMoments(iMoments_),
+  pMoments(setting_)
+{
+  // Allocation of particles
+  // part = new Particles3D[ns];
+  part = (Particles3D*) malloc(sizeof(Particles3D)*ns);
+  for (int i = 0; i < ns; i++)
+  {
+    // placement new
+    new(&part[i]) Particles3D(i,col,vct,grid);
+  }
+}
+
+/*  -------------- */
+/*!  Particle mover */
+/*  -------------- */
+bool Kinetics::moveParticles(const_arr4_double aosEMfield)
+{
+  // move all species of particles
+  {
+    timeTasks_set_main_task(TimeTasks::PARTICLES);
+
+    pad_particle_capacities();
+
+    #pragma omp parallel
+    {
+    for (int i = 0; i < ns; i++)  // move each species
+    {
+      // #pragma omp task inout(part[i]) in(grid) target_device(booster)
+      //
+      // use the Predictor Corrector scheme to move particles
+      switch(Parameters::get_MOVER_TYPE())
+      {
+        case Parameters::SoA:
+          part[i].mover_PC(aosEMfield);
+          break;
+        //case Parameters::SoA_vec_resort:
+        //  part[i].mover_PC_vectorized(aosEMfield);
+        //  break;
+        case Parameters::AoS:
+          part[i].mover_PC_AoS(aosEMfield);
+          break;
+        case Parameters::AoSintr:
+          part[i].mover_PC_AoS_vec_intr(aosEMfield);
+          break;
+        case Parameters::AoSvec:
+          part[i].mover_PC_AoS_vec(aosEMfield);
+          break;
+        //case Parameters::AoS_vec_onesort:
+        //  part[i].mover_PC_AoS_vec_onesort(aosEMfield);
+        //  break;
+        default:
+          unsupported_value_error(Parameters::get_MOVER_TYPE());
+      }
+      // overlap initial communication of electrons with moving of ions
+      #pragma omp master
+      part[i].separate_and_send_particles();
+    }
+    }
+    for (int i = 0; i < ns; i++)  // communicate each species
+    {
+      //part[i].communicate_particles();
+      part[i].recommunicate_particles_until_done(1);
+    }
+  }
+
+  /* -------------------------------------- */
+  /* Repopulate the buffer zone at the edge */
+  /* -------------------------------------- */
+
+  for (int i=0; i < ns; i++) {
+    if (col->getRHOinject(i)>0.0)
+      part[i].repopulate_particles();
+  }
+
+  /* --------------------------------------- */
+  /* Remove particles from depopulation area */
+  /* --------------------------------------- */
+
+  if (col->getCase()=="Dipole") {
+    for (int i=0; i < ns; i++)
+    {
+      double Qremoved = part[i].deleteParticlesInsideSphere( col->getL_square(),
+        col->getx_center(),col->gety_center(),col->getz_center());
+    }
+  }
+  return (false);
+}
+
+void Kinetics::sortParticles() {
+  eprintf("check how this is being used.");
+  for(int species_idx=0; species_idx<ns; species_idx++)
+    part[species_idx].sort_particles();
+}
+
+void Kinetics::pad_particle_capacities()
+{
+  for (int i = 0; i < ns; i++)
+    part[i].pad_capacities();
+}
+
+// convert particle to struct of arrays (assumed by I/O)
+void Kinetics::convertParticlesToSoA()
+{
+  eprintf("check how this is being used.");
+  for (int i = 0; i < ns; i++)
+    part[i].convertParticlesToSoA();
+}
+
+// convert particle to array of structs (used in computing)
+void Kinetics::convertParticlesToAoS()
+{
+  eprintf("check how this is being used.");
+  for (int i = 0; i < ns; i++)
+    part[i].convertParticlesToAoS();
+}
+
+// convert particle to array of structs (used in computing)
+void Kinetics::convertParticlesToSynched()
+{
+  for (int i = 0; i < ns; i++)
+    part[i].convertParticlesToSynched();
+}
+
